Fix TorrentProject provider
